Moved import logic into the options screen
